requirements.txt
aniso8601==1.1.0
click==6.6
coloredlogs==5.1.1
Flask==0.11.1
flask-restplus==0.9.2
humanfriendly==2.1
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
marshmallow==2.10.3
python-dateutil==2.5.3
pytz==2016.7
redis==2.10.5
requests==2.11.1
rethinkdb==2.3.0.post6
six==1.10.0
Werkzeug==0.11.11
